+++ src/terminalui.cpp
@@ -153,7 +153,7 @@ std::string TerminalUI::drawListMenu (const char * title, const char * message,
         case 0:
             /* F12 is pressed, and we don't care; continue to case 1 */
         case 1:
-            return items[selector];
+            return items[static_cast<unsigned long>(selector)];
         case 2:
             abortInstall();
             break;
@@ -204,7 +204,7 @@ std::vector<std::string> TerminalUI::drawFieldMenu (
 
     char** fieldEntries;
     struct newtWinEntry* field;
-    unsigned vectorSize = items.size();
+    size_t vectorSize = items.size();
 
     field = new struct newtWinEntry[vectorSize + 1]();
     fieldEntries = new char*[vectorSize + 1]();
